move CGIFHelper to separate file
