+++ src/game/client/client_base.vpc
@@ -355,6 +355,7 @@ $Project
 		$File	"$SRCDIR\game\shared\gamestringpool.cpp"
 		$File	"gametrace_client.cpp"
 		$File	"$SRCDIR\game\shared\gamevars_shared.cpp"
+		$File	"$SRCDIR\common\gifhelper.cpp"
 		$File	"glow_outline_effect.cpp"
 		$File	"glow_overlay.cpp"
 		$File	"$SRCDIR\game\shared\hintmessage.cpp"
@@ -774,6 +775,7 @@ $Project
 		$File	"fx_trail.h"
 		$File	"fx_water.h"
 		$File	"$SRCDIR\game\shared\GameEventListener.h"
+		$File	"$SRCDIR\common\gifhelper.h"
 		$File	"glow_outline_effect.h"
 		$File	"glow_overlay.h"
 		$File	"$SRCDIR\game\shared\hintmessage.h"
