+++ src/game/client/vgui_avatarimage.h
@@ -14,8 +14,7 @@
 #include <vgui_controls/ImagePanel.h>
 #include "steam/steam_api.h"
 #include "c_baseplayer.h"
-#include "utlarray.h"
-#include "utlpair.h"
+#include "gifhelper.h"
 
 // size of the friend background frame (see texture ico_friend_indicator_avatar)
 #define FRIEND_ICON_SIZE_X	(55)	
@@ -27,41 +26,6 @@
 
 // size of the standard avatar icon (unless override by SetAvatarSize)
 #define DEFAULT_AVATAR_SIZE		(32)
-
-typedef struct GifFileType GifFileType;
-typedef unsigned char GifByteType;
-
-//-----------------------------------------------------------------------------
-// Purpose: Simple utility for decoding GIFs
-//-----------------------------------------------------------------------------
-class CGIFHelper
-{
-public:
-	CGIFHelper( void ) : m_pImage( NULL ), m_pPrevFrameBuffer( NULL ),
-		m_iSelectedFrame( 0 ), m_dIterateTime( 0.0 ) {}
-	~CGIFHelper( void ) { CloseImage(); }
-
-	bool OpenImage( CUtlBuffer* pBuf );
-	void CloseImage( void );
-
-	// iterates to the next frame, returns true if we have just looped
-	bool NextFrame( void );
-	int GetSelectedFrame( void ) const { return m_iSelectedFrame; }
-	bool ShouldIterateFrame( void ) const { return m_dIterateTime < Plat_FloatTime(); }
-
-	// retrieve data for the current frame
-	void GetRGBA( uint8** ppOutFrameBuffer ); // size of the out frame buffer should be iScreenWide * iScreenTall * 4
-	void GetFrameSize( int& iWide, int& iTall ) const;
-	void GetScreenSize( int& iWide, int& iTall ) const;
-
-private:
-	static int ReadData( GifFileType* pFile, GifByteType* pBuffer, int cubBuffer );
-
-	GifFileType* m_pImage;
-	uint8* m_pPrevFrameBuffer;
-	int m_iSelectedFrame;
-	double m_dIterateTime;
-};
 
 struct AnimatedAvatar_t
 {
